create app directory, move App.js App.css
