+++ stock-tracker/src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import 'bootstrap/dist/css/bootstrap.min.css';
-import App from './App';
+import App from './app/App';
 import { BrowserRouter } from 'react-router-dom';
 
 ReactDOM.render(
